training/requirements.txt
# AI/ML Dependencies for Cheat Detection Training
opencv-python==4.8.1.78
mediapipe==0.10.7
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tensorflowjs==4.11.0

# Additional utilities
tqdm==4.66.1
Pillow==10.0.0
jupyter==1.0.0
notebook==7.0.2